Added Check_RepositoryFactory_Pattern, Check_RepositoryFactory_Patterns, Check_RepositoryFactory_PatternDto, Check_RepositoryFactory_PatternDtos unit-tests. Some renamings.
